% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette_mono.R
\name{color_palette_mono}
\alias{color_palette_mono}
\title{Generate colors for monotherapy experiments}
\usage{
color_palette_mono(joined_data_frame, category = "treatment_name")
}
\arguments{
\item{joined_data_frame}{A data frame object created from the data wrangling
RShiny}

\item{category}{The variable used for determining where each condition is
located. Values include "treatment_name" and "well_annotation". Defaults to
"well_annotation"}
}
\value{
A data frame with the variable "color" for the color pallete
}
\description{
Generate colors for monotherapy experiments
}
\examples{
data_frame_colors <- color_palette_mono(joined_data_frame = data_frame)
}
