% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_dss_barplot.R
\name{ctg_dss_barplot}
\alias{ctg_dss_barplot}
\title{DSS3 Bar Chart}
\usage{
ctg_dss_barplot(
  ctg_list,
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  y_axis_title_size = 14,
  score_label_size = 4,
  plot_title_size = 20
)
}
\arguments{
\item{ctg_list}{A list object output from the dr4pl_fit_loop function}

\item{x_axis_text_size}{A numeric value specifying the size of the
x-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_text_size}{A numeric value specifying the size of the
y-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_title_size}{A numeric value specifying the size of the
y-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{score_label_size}{A numeric value specifying the size of the
score labels above the bars. Defaults to 4.}

\item{plot_title_size}{A numeric value specifying the size of the plot
title. Defaults to 20. See \code{\link[ggplot2]{element_text}}.}
}
\value{
A ggplot2 bar chart
}
\description{
DSS3 Bar Chart
}
