% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_ic50_forest_plot.R
\name{ctg_ic50_forest_plot}
\alias{ctg_ic50_forest_plot}
\title{CTG IC50 Forest Plot}
\usage{
ctg_ic50_forest_plot(
  ctg_list,
  x_scale = "log",
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  x_axis_title_size = 14,
  n_x_axis_breaks = 8,
  plot_title_size = 20
)
}
\arguments{
\item{ctg_list}{A list object output from the d4pl_fit_loop function}

\item{x_scale}{A character string specifying whether the x-axis should be
transformed. Values include "log" for log-transformed concentration and
"standard" for non-transformed concentration scale. Defaults to "log".}

\item{x_axis_text_size}{A numeric value specifying the size of the
x-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_text_size}{A numeric value specifying the size of the
y-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{x_axis_title_size}{A numeric value specifying the size of the
x-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{n_x_axis_breaks}{An integer specifying the number of major breaks
for the x-axis. Defaults to 6.}

\item{plot_title_size}{A numeric value specifying the size of the plot
title. Defaults to 20. See \code{\link[ggplot2]{element_text}}.}
}
\value{
A ggplot2 object
}
\description{
CTG IC50 Forest Plot
}
