% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_normalize.R
\name{ctg_normalize}
\alias{ctg_normalize}
\title{Normalize CTG dataa}
\usage{
ctg_normalize(data_frame, use_positive_control = TRUE)
}
\arguments{
\item{data_frame}{A data frame of CTG data}

\item{use_positive_control}{A logical value specifying whether to normalize
by the postive control as well. Defaults to TRUE.}
}
\value{
A data frame object with normalized ctg data
}
\description{
Normalize CTG dataa
}
