% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_qc_control_plot.R
\name{ctg_qc_control_plot}
\alias{ctg_qc_control_plot}
\title{CTG QC Control Plot}
\usage{
ctg_qc_control_plot(
  data_frame,
  show_outlier = TRUE,
  make_interactive = FALSE,
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  plot_title_size = 20,
  geom_point_size = 2.5
)
}
\arguments{
\item{data_frame}{A data frame of CTG data}

\item{show_outlier}{A logical value specifying whether to show the points
marked as outliers/ Defaults to TRUE.}

\item{make_interactive}{A logical value specifying whether to convert the
plot into an interactive plotly object. Defaults to FALSE.}

\item{x_axis_text_size}{A numeric value specifying the size of the
x-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_text_size}{A numeric value specifying the size of the
y-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{x_axis_title_size}{A numeric value specifying the size of the
x-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_title_size}{A numeric value specifying the size of the
y-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{plot_title_size}{A numeric value specifying the size of the plot
title. Defaults to 20. See \code{\link[ggplot2]{element_text}}.}

\item{geom_point_size}{A numeric value specifying the size of the points
on the graph. Defaults to 2. See \code{\link[ggplot2]{geom_point}}.}
}
\value{
Returns a ggplot2 plot of the input data. When make_interactive is
TRUE, returns an interactive plotly object version of the ggplot2 plot.
}
\description{
CTG QC Control Plot
}
