% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_qc_mean_outlier.R
\name{ctg_qc_mean_outlier}
\alias{ctg_qc_mean_outlier}
\title{Detect outlier with z-scores (mean/sd)}
\usage{
ctg_qc_mean_outlier(ctg_data, z_score_threshold = 3)
}
\arguments{
\item{ctg_data}{A list object (ctg_list) or data frame with CTG data}

\item{z_score_threshold}{A numeric value specifying the z-score threshold to
use for determining outliers. Defaults to 3.}
}
\value{
A data frame object
}
\description{
Detect outlier with z-scores (mean/sd)
}
