% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_qc_output.R
\name{ctg_qc_output}
\alias{ctg_qc_output}
\title{Create QC output list for growth datasets}
\usage{
ctg_qc_output(ctg_list, outlier_manual_only = TRUE, prism = TRUE)
}
\arguments{
\item{ctg_list}{A list object containing ctg data}

\item{outlier_manual_only}{A logical value that when TRUE filters outliers in
terms of the outlier_manual_yn variable, if FALSE is will consider outliers
as those annotated by both the outlier_manual_yn and the outlier_auto_yn.
Defaults to TRUE.}

\item{prism}{A logical value specifying whether to add a dataset
for importing into GraphPad PRISM. Defaults to TRUE.}
}
\value{
A list object
}
\description{
Create QC output list for growth datasets
}
