% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_qc_treat_plot.R
\name{ctg_qc_treat_plot}
\alias{ctg_qc_treat_plot}
\title{CTG QC Treatment Plot}
\usage{
ctg_qc_treat_plot(
  ctg_list,
  treat_name = "",
  show_outlier = TRUE,
  make_interactive = FALSE,
  x_limits = c(NA, NA),
  x_axis_breaks = ggplot2::waiver(),
  n_x_axis_breaks = 6,
  y_limits = c(NA, NA),
  y_axis_breaks = ggplot2::waiver(),
  n_y_axis_breaks = NULL,
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  plot_title_size = 20,
  geom_point_size = 2.5,
  geom_line_width = 1
)
}
\arguments{
\item{ctg_list}{A list object generated from dr4pl_qc_fit_loop}

\item{treat_name}{A character string specifying the treatment_name to filter
on.}

\item{show_outlier}{A logical value specifying whether to show the points
marked as outliers/ Defaults to TRUE.}

\item{make_interactive}{A logical value specifying whether to convert the
plot into an interactive plotly object. Defaults to FALSE.}

\item{x_limits}{A numeric vector of length two providing limits for
the x-axis. Use NA to refer to the existing minimum or maximum.
Defaults to c(NA, NA). See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{x_axis_breaks}{A numeric vector specifying manual numeric breaks.
Defaults to ggplot2::waiver(). See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{n_x_axis_breaks}{An integer specifying the number of major breaks
for the x-axis. Defaults to 6.}

\item{y_limits}{A numeric vector of length two providing limits for
the y-axis. Use NA to refer to the existing minimum or maximum.
Defaults to c(NA, NA). See \code{\link[ggplot2]{scale_y_continuous}}.}

\item{y_axis_breaks}{A numeric vector specifying manual numeric breaks.
Defaults to ggplot2::waiver(). See \code{\link[ggplot2]{scale_y_continuous}}.}

\item{n_y_axis_breaks}{An integer specifying the number of major breaks
for the x-axis. Defaults to NULL.}

\item{x_axis_text_size}{A numeric value specifying the size of the
x-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_text_size}{A numeric value specifying the size of the
y-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{x_axis_title_size}{A numeric value specifying the size of the
x-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_title_size}{A numeric value specifying the size of the
y-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{plot_title_size}{A numeric value specifying the size of the plot
title. Defaults to 20. See \code{\link[ggplot2]{element_text}}.}

\item{geom_point_size}{A numeric value specifying the size of the points
on the graph. Defaults to 2.5. See \code{\link[ggplot2]{geom_point}}.}

\item{geom_line_width}{A numeric value specifying the width of the line.
Defaults to 1.}
}
\value{
Returns a ggplot2 plot of the input data. When make_interactive is
TRUE, returns an interactive plotly object version of the ggplot2 plot.
}
\description{
CTG QC Treatment Plot
}
