% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_summary_table_all.R
\name{ctg_summary_table_all}
\alias{ctg_summary_table_all}
\title{CTG Summary Table All Treatments}
\usage{
ctg_summary_table_all(
  ctg_list,
  show = "both",
  font_name = "Calibri",
  font_size_header = 14,
  font_size_body = 12,
  use_knit_print = FALSE
)
}
\arguments{
\item{ctg_list}{A list object generated by the dr4pl_fit_loop function}

\item{show}{A character string specifying what kind of sub titles to
display on the plat. Values include "ic50" for the
relative IC50, "dss3" for the drug sensitivity score, or "both" for
displaying the relative IC50 and the DSS3 score. Defaults to "both".}

\item{font_name}{A character string specifying the name of the font to use
when rendering the table. Defaults to "Albany AMT".
See \code{\link[flextable]{font}}.}

\item{font_size_header}{A numeric value specifying the size of the font
for the header of the table. Defaults to 14.
See \code{\link[flextable]{fontsize}}.}

\item{font_size_body}{A numeric value specifying the size of the font
for the body of the table. Defaults to 12.
See \code{\link[flextable]{fontsize}}.}

\item{use_knit_print}{A logical value to specify whether the flextable
should be printed using \code{\link[knitr]{knit_print}} function
instead of the flextable object being returned. Defaults to FALSE.}
}
\value{
A flextable object
}
\description{
CTG Summary Table All Treatments
}
