% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_summary_table_ind.R
\name{ctg_summary_table_ind}
\alias{ctg_summary_table_ind}
\title{CTG Summary Table All Treatments}
\usage{
ctg_summary_table_ind(
  ctg_list,
  treat_name = "",
  font_name = "Calibri",
  font_size_header = 14,
  font_size_body = 12,
  use_knit_print = FALSE
)
}
\arguments{
\item{ctg_list}{A list object generated by the dr4pl_fit_loop function}

\item{treat_name}{A character string specifying the treatment_name to
filter on.}

\item{font_name}{A character string specifying the name of the font to use
when rendering the table. Defaults to "Albany AMT".
See \code{\link[flextable]{font}}.}

\item{font_size_header}{A numeric value specifying the size of the font
for the header of the table. Defaults to 14.
See \code{\link[flextable]{fontsize}}.}

\item{font_size_body}{A numeric value specifying the size of the font
for the body of the table. Defaults to 12.
See \code{\link[flextable]{fontsize}}.}

\item{use_knit_print}{A logical value to specify whether the flextable
should be printed using \code{\link[knitr]{knit_print}} function
instead of the flextable object being returned. Defaults to FALSE.}
}
\value{
A flextable object
}
\description{
CTG Summary Table All Treatments
}
