% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_to_prism.R
\name{ctg_to_prism}
\alias{ctg_to_prism}
\title{Convert ctg dataset to PRISM shape}
\usage{
ctg_to_prism(data_frame)
}
\arguments{
\item{data_frame}{A ctg data frame object}
}
\value{
A dataframe that is shaped for copying and pasting into PRISM
}
\description{
Convert ctg dataset to PRISM shape
}
