% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_treat_plot_all.R
\name{ctg_treat_plot_all}
\alias{ctg_treat_plot_all}
\title{Create Dose-Response Curve Plot for All Treatments}
\usage{
ctg_treat_plot_all(
  ctg_list,
  x_scale = "log",
  display_type = "points",
  y_axis_title = "Normalized Value",
  make_interactive = FALSE,
  x_limits = c(NA, NA),
  x_axis_breaks = ggplot2::waiver(),
  n_x_axis_breaks = 6,
  y_limits = c(NA, NA),
  y_axis_breaks = ggplot2::waiver(),
  n_y_axis_breaks = 7,
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  plot_title_size = 18,
  plot_sub_title_size = 16,
  geom_point_size = 2.5,
  geom_line_width = 1
)
}
\arguments{
\item{ctg_list}{A list object generated by the dr4pl_git_loop function}

\item{x_scale}{A character string specifying whether the x-axis should be
transformed. Values include "log" for log-transformed concentration and
"standard" for non-transformed concentration scale. Defaults to "log".}

\item{display_type}{A character string specifying whether to display the
normalized values "points" or the mean+/-se bars "se_bars". Defaults to
"points"}

\item{make_interactive}{A logical value specifying whether to convert the
plot into an interactive plotly object. Defaults to FALSE.}

\item{x_limits}{A numeric vector of length two providing limits for
the x-axis. Use NA to refer to the existing minimum or maximum.
Defaults to c(NA, NA). See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{x_axis_breaks}{A numeric vector specifying manual numeric breaks.
Defaults to ggplot2::waiver(). See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{n_x_axis_breaks}{An integer specifying the number of major breaks
for the x-axis. Defaults to 6.}

\item{y_limits}{A numeric vector of length two providing limits for
the y-axis. Use NA to refer to the existing minimum or maximum.
Defaults to c(NA, NA). See \code{\link[ggplot2]{scale_y_continuous}}.}

\item{y_axis_breaks}{A numeric vector specifying manual numeric breaks.
Defaults to ggplot2::waiver(). See \code{\link[ggplot2]{scale_y_continuous}}.}

\item{n_y_axis_breaks}{An integer specifying the number of major breaks
for the x-axis. Defaults to NULL.}

\item{x_axis_text_size}{A numeric value specifying the size of the
x-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_text_size}{A numeric value specifying the size of the
y-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{x_axis_title_size}{A numeric value specifying the size of the
x-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_title_size}{A numeric value specifying the size of the
y-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{plot_title_size}{A numeric value specifying the size of the plot
title. Defaults to 18. See \code{\link[ggplot2]{element_text}}.}

\item{plot_sub_title_size}{A numeric value specifying the size of the plot
title. Defaults to 16. See \code{\link[ggplot2]{element_text}}.}

\item{geom_point_size}{A numeric value specifying the size of the points
on the graph. Defaults to 2.5. See \code{\link[ggplot2]{geom_point}}.}

\item{geom_line_width}{A numeric value specifying the width of the line.
Defaults to 1.}
}
\value{
A ggplot2 or plotly object
}
\description{
Create Dose-Response Curve Plot for All Treatments
}
