% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_treat_plot_ind.R
\name{ctg_treat_plot_ind}
\alias{ctg_treat_plot_ind}
\title{Create CTG Dose-Response Curve Plot}
\usage{
ctg_treat_plot_ind(
  ctg_list,
  treat_name = "",
  x_scale = "log",
  display_type = "points",
  sub_title = "none",
  y_axis_title = "Normalized Value",
  make_interactive = FALSE,
  x_limits = c(NA, NA),
  x_axis_breaks = ggplot2::waiver(),
  n_x_axis_breaks = 6,
  y_limits = c(NA, NA),
  y_axis_breaks = ggplot2::waiver(),
  n_y_axis_breaks = NULL,
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  plot_title_size = 18,
  geom_point_size = 2.5,
  geom_line_width = 1
)
}
\arguments{
\item{ctg_list}{A Description}

\item{treat_name}{A Description}

\item{x_scale}{A Description}

\item{display_type}{A Description}

\item{sub_title}{A Description}

\item{y_axis_title}{A Description}

\item{make_interactive}{A Description}

\item{x_limits}{A Description}

\item{x_axis_breaks}{A Description}

\item{n_x_axis_breaks}{A Description}

\item{y_limits}{A Description}

\item{y_axis_breaks}{A Description}

\item{n_y_axis_breaks}{A Description}

\item{x_axis_text_size}{A Description}

\item{y_axis_text_size}{A Description}

\item{x_axis_title_size}{A Description}

\item{y_axis_title_size}{A Description}

\item{plot_title_size}{A Description}

\item{geom_point_size}{A Description}

\item{geom_line_width}{A Description}
}
\value{
A ggplot2 object or a plotly object
}
\description{
Create CTG Dose-Response Curve Plot
}
