% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr4pl_fit_loop.R
\name{dr4pl_fit_loop}
\alias{dr4pl_fit_loop}
\title{DR4PL Curve Fitting Loop for Data Analysis}
\usage{
dr4pl_fit_loop(
  data_frame,
  concentration_unit = "uM",
  method_init = "logistic",
  method_robust = "squared",
  lb_if_min_gt = 0.3,
  ub_if_max_lt = 0.8,
  readout = "activity",
  activity_threshold = 10,
  dss_type = 3,
  log_transform = TRUE,
  slope_threshold = 0.1,
  max_dss = 100
)
}
\arguments{
\item{data_frame}{A data frame with ctg monotherapy treatment data}

\item{concentration_unit}{A character string specifying the concentration
units. Defaults to uM.}

\item{method_init}{Method of obtaining initial values of the parameters.
If this parameter is left unassigned, a default "Mead" method will be used.
Assign "logistic" to use the logistic method.}

\item{method_robust}{Parameter to select loss function for the robust
estimation method to be used to fit a model. The argument NULL indicates
the sum of squares loss, "absolute" indicates the absolute deviation loss,
"Huber" indicates Huber's loss and "Tukey" indicates Tukey's biweight loss.}

\item{lb_if_min_gt}{Threshold to fit the model using boundaries if the
minimum normalized value is greater than the designated threshold. Defaults to
0.3.}

\item{ub_if_max_lt}{Threshold to fit the model using boundaries if the
maximum normalized value is less than the designated threshold. Defaults to
0.8.}

\item{readout}{A character string specifying the readout type. Values include
"inhibition", "activity". Defaults to "inhibition".}

\item{activity_threshold}{A numeric value specifying the minimim actvity
(in terms of \%inhibition) needed to be counted towards the DSS. Defaults to
10.}

\item{dss_type}{A numeric value specifying which DSS score to return. Values
include 1, 2, 3. Defaults to 3.}

\item{log_transform}{A logical value specifying whether the min and max
concentrations need to be log transformed before hand. Defaults to TRUE.}

\item{slope_threshold}{A numeric value specifying at what slope is too small
and sets the DSS score to 0. Defaults to 0.1.}

\item{max_dss}{A numeric value specying the maximum a DSS score van be,
defaults to 100.}
}
\value{
A list object to be used for plotting, displaying tables, and
exporting
}
\description{
DR4PL Curve Fitting Loop for Data Analysis
}
