% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr4pl_qc_fit_loop.R
\name{dr4pl_qc_fit_loop}
\alias{dr4pl_qc_fit_loop}
\title{DR4PL Curve Fitting Loop for CTG QC}
\usage{
dr4pl_qc_fit_loop(
  data_frame,
  concentration_unit = "uM",
  method_init = "logistic",
  method_robust = "Huber",
  lb_if_min_gt = 0.3,
  ub_if_max_lt = 0.8
)
}
\arguments{
\item{data_frame}{A data frame with CTG data}

\item{concentration_unit}{A character string specifying the concentration
units. Defaults to uM.}

\item{method_init}{Method of obtaining initial values of the parameters.
If this parameter is left unassigned, a default "Mead" method will be used.
Assign "logistic" to use the logistic method.}

\item{method_robust}{Parameter to select loss function for the robust
estimation method to be used to fit a model. The argument NULL indicates
the sum of squares loss, "absolute" indicates the absolute deviation loss,
"Huber" indicates Huber's loss and "Tukey" indicates Tukey's biweight loss.}

\item{lb_if_min_gt}{Threshold to fit the model using boundaries if the
minimum normalized value is greater than the designated threshold. Defaults to
0.3.}

\item{ub_if_max_lt}{Threshold to fit the model using boundaries if the
maximum normalized value is less than the designated threshold. Defaults to
0.8.}
}
\value{
A list object
}
\description{
DR4PL Curve Fitting Loop for CTG QC
}
