% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dss_calculator.R
\name{dss_calculator}
\alias{dss_calculator}
\title{Drug Sensitivity Score Calculation}
\usage{
dss_calculator(
  data_frame,
  readout = "inhibition",
  activity_threshold = 10,
  dss_type = 3,
  concentration_unit = "uM",
  log_transform = TRUE,
  slope_threshold = 0.1,
  max_dss = 100
)
}
\arguments{
\item{data_frame}{A data frame of the 4-parameter model output within the
dr4pl_fit function}

\item{readout}{A character string specifying the readout type. Values include
"inhibition", "activity". Defaults to "inhibition".}

\item{activity_threshold}{A numeric value specifying the minimim actvity
(in terms of \%inhibition) needed to be counted towards the DSS. Defaults to
10.}

\item{dss_type}{A numeric value specifying which DSS score to return. Values
include 1, 2, 3. Defaults to 3.}

\item{concentration_unit}{A character string specying the concentration unit
needed to convert to molar. Values include "M", "mM", "uM", "nM", "fM".
Defaults to "uM"}

\item{log_transform}{A logical value specifying whether the min and max
concentrations need to be log transformed before hand. Defaults to TRUE.}

\item{slope_threshold}{A numeric value specifying at what slope is too small
and sets the DSS score to 0. Defaults to 0.1.}

\item{max_dss}{A numeric value specying the maximum a DSS score van be,
defaults to 100.}
}
\value{
A numeric value, the Drug Sensitivity Score
}
\description{
Drug Sensitivity Score Calculation
}
