% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_analysis_control_plot.R
\name{growth_analysis_control_plot}
\alias{growth_analysis_control_plot}
\title{Growth Analysis Controls Plot}
\usage{
growth_analysis_control_plot(
  data_frame,
  treatment_name = "",
  display_metric = "wells",
  make_interactive = FALSE,
  growth_metric_name = "growth_metric",
  time_units = "hours",
  concentration_units = "µM",
  x_limits = c(NA, NA),
  x_axis_breaks = ggplot2::waiver(),
  n_x_axis_breaks = NULL,
  y_limits = c(NA, NA),
  y_axis_breaks = ggplot2::waiver(),
  n_y_axis_breaks = NULL,
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  plot_title_size = 20,
  geom_point_size = 2,
  geom_line_width = 0.5
)
}
\arguments{
\item{data_frame}{A data frame object for growth based data}

\item{treatment_name}{A character string specifying the name of the current
treatment. Defaults to "".}

\item{display_metric}{A character string specifying the type of points to
display. Values include "wells" for well-level data, "mean_se" for displaying
the mean points and se bars, "mean" for just the mean of the points, or
"median" for the median of the points. Defaults to "well".}

\item{make_interactive}{A logical value specifying whether to convert the
plot into an interactive plotly object. Defaults to FALSE.}

\item{growth_metric_name}{A character string for specifying the name of
the growth metric (y-axis title) to be displayed on the plot.
Defaults to "growth_metric".}

\item{time_units}{A character string for specifying the name of the time
variable (x-axis title) to be displayed on the plot. Defaults to "hours".}

\item{concentration_units}{A character string specifying the concetration
units. Defaults to "µM".}

\item{x_limits}{A numeric vector of length two providing limits for
the x-axis. Use NA to refer to the existing minimum or maximum.
Defaults to c(NA, NA). See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{x_axis_breaks}{A numeric vector specifying manual numeric breaks.
Defaults to ggplot2::waiver(). See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{n_x_axis_breaks}{An integer specifying the number of major breaks
for the x-axis. Defaults to NULL.}

\item{y_limits}{A numeric vector of length two providing limits for
the y-axis. Use NA to refer to the existing minimum or maximum.
Defaults to c(NA, NA). See \code{\link[ggplot2]{scale_y_continuous}}.}

\item{y_axis_breaks}{A numeric vector specifying manual numeric breaks.
Defaults to ggplot2::waiver(). See \code{\link[ggplot2]{scale_y_continuous}}.}

\item{n_y_axis_breaks}{An integer specifying the number of major breaks
for the x-axis. Defaults to NULL.}

\item{x_axis_text_size}{A numeric value specifying the size of the
x-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_text_size}{A numeric value specifying the size of the
y-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{x_axis_title_size}{A numeric value specifying the size of the
x-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_title_size}{A numeric value specifying the size of the
y-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{plot_title_size}{A numeric value specifying the size of the plot
title. Defaults to 20. See \code{\link[ggplot2]{element_text}}.}

\item{geom_point_size}{A numeric value specifying the size of the points
on the graph. Defaults to 2. See \code{\link[ggplot2]{geom_point}}.}

\item{geom_line_width}{A numeric value specifying the width of the line.
Defaults to 0.5.}
}
\value{
A ggplot2 or plotly object
}
\description{
Growth Analysis Controls Plot
}
