% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_data_prep.R
\name{growth_data_prep}
\alias{growth_data_prep}
\title{Import and prep Incucyte/Cytation data}
\usage{
growth_data_prep(file_path = "", imaging_equipment = "Incucyte")
}
\arguments{
\item{file_path}{A character string specifying the full file path to the
Incucyte/Cytation text file (inclusive of the full growth .txt file name)}

\item{imaging_equipment}{A character string specifying the machine in which
the file was generated from. Values include "Incucyte" or "Cytation".
Defaults to Cytation}
}
\value{
A data frame of the prepared growth data
}
\description{
Import and prep Incucyte/Cytation data
}
\examples{
growth_data_prep(file_path = "",
imaging_equipment = "Incucyte")
}
