% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_qc_output.R
\name{growth_qc_output}
\alias{growth_qc_output}
\title{Create QC output list for growth datasets}
\usage{
growth_qc_output(
  data_frame,
  outlier_manual_only = TRUE,
  growthcurveme = TRUE,
  lgrscore = TRUE,
  prism = TRUE
)
}
\arguments{
\item{data_frame}{A growth data frame object.}

\item{outlier_manual_only}{A logical value that when TRUE filters outliers in
terms of the outlier_manual_yn variable, if FALSE is will consider outliers
as those annotated by both the outlier_manual_yn and the outlier_auto_yn.
Defaults to TRUE.}

\item{growthcurveme}{A logical value specifying whether to add a dataset
for importing into the GrowthCurveME RShiny. Defaults to TRUE.}

\item{lgrscore}{A logical value specifying whether to add a dataset
for importing into the LGRscore RShiny. Defaults to TRUE.}

\item{prism}{A logical value specifying whether to add a dataset
for importing into GraphPad PRISM. Defaults to TRUE.}
}
\value{
A list object
}
\description{
Create QC output list for growth datasets
}
