% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_to_prism.R
\name{growth_to_prism}
\alias{growth_to_prism}
\title{Convert growth dataset to PRISM shape}
\usage{
growth_to_prism(data_frame)
}
\arguments{
\item{data_frame}{A growth-based data frame}
}
\value{
A dataframe object shaped in preparation for importing into PRISM
}
\description{
Convert growth dataset to PRISM shape
}
