% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labguru_plate_prep.R
\name{labguru_plate_prep}
\alias{labguru_plate_prep}
\title{Import and prep Labguru plate map}
\usage{
labguru_plate_prep(file_path = "")
}
\arguments{
\item{file_path}{A character string specifying the full file path to the
Labguru plate map file (inclusive of the plate map excel file name)}
}
\value{
A data frame of the Labguru plate map
}
\description{
Import and prep Labguru plate map
}
\examples{
labguru_plate_prep(file_path = "")
}
