% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_outlier_fit.R
\name{loess_outlier_fit}
\alias{loess_outlier_fit}
\title{Loess outlier detection}
\usage{
loess_outlier_fit(data_frame, span_value = 0.3, residual_threshold = 3)
}
\arguments{
\item{data_frame}{A data frame with at minimum the columns well, time, and
growth_metric}

\item{span_value}{A LOESS regression parameter that controls the level of
smoothing that occurs across the data. Defaults to 0.3.}

\item{residual_threshold}{A threshold for determining the absolute value
of a standardized residual from the loess regression to mark as an outlier.
Defaults to 3.}
}
\value{
The input data frame with additional columns for the predicted
growth_metric, residuals, and outlier_auto_yn
}
\description{
Loess outlier detection
}
\examples{
data_clean <- loess_outlier_fit(data_frame, span_value = 0.3)
}
