% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_decimals.R
\name{max_decimals}
\alias{max_decimals}
\title{Calculate max decimals}
\usage{
max_decimals(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeriv value, the maximum number of decimal points
}
\description{
Calculate max decimals
}
