% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_data_join.R
\name{plate_data_join}
\alias{plate_data_join}
\title{Join multiple prepared datasets together}
\usage{
plate_data_join(
  labguru_plate_data_frame = NULL,
  tecan_plate_data_frame = NULL,
  growth_data_frame = NULL,
  ctg_data_frame = NULL
)
}
\arguments{
\item{labguru_plate_data_frame}{A data frame of a prepared labguru plate map
generated by the \code{\link{labguru_plate_prep}} function}

\item{tecan_plate_data_frame}{A data frame of a prepared Tecan report
generated by the \code{\link{labguru_plate_prep}} function}

\item{growth_data_frame}{A data frame of a prepared Growth file from
Incucyte/Cytation generated by the \code{\link{growth_data_prep}} function}

\item{ctg_data_frame}{A data frame of a prepared CTG file from
generated by the \code{\link{ctg_prep}} function}
}
\value{
A data frame of the respectively joined data files
}
\description{
Join multiple prepared datasets together
}
\examples{
plate_data_join(labguru_plate_data_frame = NULL,
tecan_plate_data_frame = NULL,
growth_data_frame = NULL,
ctg_data_frame = NULL)
}
