% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_report_prep.R
\name{tecan_report_prep}
\alias{tecan_report_prep}
\title{Import and prep tecan report file}
\usage{
tecan_report_prep(
  file_path = "",
  drugging_type = "Monotherapy",
  remove_na = TRUE
)
}
\arguments{
\item{file_path}{A character string specifying the full file path to the
Tecan report file (inclusive of the Tecan report .xlsx file name)}

\item{drugging_type}{The drugging type scheme of the Tecan file. Values
include "Monotherapy" and "Synergy". Defaults to "Monotherapy"}

\item{remove_na}{A logical value specifying whether to remove non-dispensed
wells (i.e. treatment_name is NA). Defaults to TRUE.}
}
\value{
A data frame of the prepared tecan report data
}
\description{
Import and prep tecan report file
}
\examples{
tecan_report_prep(
file_path = "",
drugging_type = "Monotherapy",
remove_na = TRUE)
}
\seealso{
\code{\link{tecan_report_prep_mono}}
\code{\link{tecan_report_prep_syn}}
}
