% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_report_prep_syn.R
\name{tecan_report_prep_syn}
\alias{tecan_report_prep_syn}
\title{Import and prep tecan report file - synergy}
\usage{
tecan_report_prep_syn(file_path = "", concentration_units = "micromolar")
}
\arguments{
\item{file_path}{A character string specifying the full file path to the
Tecan report file for synergy-based experiments (inclusive of the tecan
report excel file name)}

\item{concentration_units}{A character string specifying the concentration
units. Values incluide "molar", "millimolar", "micromolar", "nonomolar",
and "picomolar". Defaults to "micromolar".}
}
\value{
A data frame of the prepared tecan report data
}
\description{
Import and prep tecan report file - synergy
}
\examples{
tecan_report_prep_syn(
file_path = "",
concentration_units = "micromolar")
}
