% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_annotate_transfer.R
\name{well_annotate_transfer}
\alias{well_annotate_transfer}
\title{Transfer well_annotations}
\usage{
well_annotate_transfer(joined_data_frame)
}
\arguments{
\item{joined_data_frame}{The joined data frame created by the data wrangling
R Shiny when the Labguru plate map has been imported}
}
\value{
The joined data frame
}
\description{
Transfer well_annotations
}
\examples{
joined_data <- well_annotate_transfer(joined_data_frame = joined_data_frame)
}
