% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg_qc_mean_outlier.R
\name{ctg_qc_mean_outlier}
\alias{ctg_qc_mean_outlier}
\title{Detect outlier with z-scores (mean/sd)}
\usage{
ctg_qc_mean_outlier(data_frame, z_score_threshold = 3)
}
\arguments{
\item{data_frame}{A data frame with variable treatment_name, concentration,
and ctg_value}

\item{z_score_threshold}{A numeric value specifying the z-score threshold to
use for determining outliers. Defaults to 3.}
}
\value{
A data frame object
}
\description{
Detect outlier with z-scores (mean/sd)
}
