% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_report_prep_syn.R
\name{tecan_report_prep_syn}
\alias{tecan_report_prep_syn}
\title{Import and prep tecan report file - synergy}
\usage{
tecan_report_prep_syn(file_path = "", remove_na = TRUE)
}
\arguments{
\item{file_path}{A character string specifying the full file path to the
Tecan report file for synergy-based experiments (inclusive of the Tecan
report .xlsx file name)}

\item{remove_na}{A logical value specifying whether to remove non-dispensed
wells (i.e. treatment_name is NA). Defaults to TRUE}
}
\value{
A data frame of the prepared tecan report data
}
\description{
Import and prep tecan report file - synergy
}
\examples{
tecan_report_prep_syn(
file_path = "")
}
